Refactor the dataapi server v2 file into modular files
